requirements.txt
spotipy
atproto
python-dotenv